pyproject.toml
[project]
name = "iieleveblabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi>=0.115", "uvicorn[standard]>=0.30", "httpx>=0.27", "elevenlabs>=1.0", "python-dotenv>=1.0"]
